Add a sanity check on the student's application term with creating an assignment.
It looks like we were chekecking the student type field on assignmnet, and usually
if a student type was bad/missing, then so was the application term. Now, it looks
like we're seeing some new behavior where the student type field can be valid, but
the application term is still empty/invalid. So, now we need to specifically check
both fields. Not sure why we haven't run into this sooner.